# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Xya-StyleMind Backend API
  version: 0.1.0
  description: |
    OpenAPI contract for the Xya-StyleMind Go/Gin backend.

    All JSON API endpoints use the shared response shape:
    `{ success, message, data, meta }`.

    Authentication uses Bearer JWT access tokens. Admin endpoints require an
    authenticated user with role `admin`.
servers:
  - url: http://localhost:8080
    description: Local backend
  - url: http://localhost:8080/api/v1
    description: Local API v1 base path
...
security: []
paths:
...
  /api/v1/orders:
    post:
      tags: [Orders]
      summary: Checkout current cart into an order
      operationId: checkout
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CheckoutRequest'
            examples:
              default:
                value:
                  recipient_name: Nguyen Van A
                  phone: '0901234567'
                  address_line: 123 Nguyen Trai
                  city: Ho Chi Minh City
                  district: District 1
                  note: Call before delivery
                  shipping_method: standard
                  payment_method: cod
      responses:
        '201':
          description: Order created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderResponseEnvelope'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalServerError'
    get:
      tags: [Orders]
      summary: List current user's orders
      operationId: listMyOrders
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/LimitParam'
      responses:
        '200':
          description: User orders
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderListResponseEnvelope'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
  /api/v1/orders/{id}/return-requests:
    post:
      tags: [Returns]
      summary: Request a return or refund for an order
      description: Authenticated users can request returns for their own paid, shipping, or completed orders.
      operationId: createReturnRequest
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/OrderIDPath'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateReturnRequest'
            examples:
              default:
                value:
                  reason: The size does not fit and I would like a refund.
      responses:
        '201':
          description: Return request created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReturnRequestResponseEnvelope'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
  /api/v1/me/return-requests:
    get:
      tags: [Returns]
      summary: List current user's return requests
      operationId: listMyReturnRequests
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/LimitParam'
      responses:
        '200':
          description: User return requests
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReturnRequestListResponseEnvelope'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    PageParam:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
      example: 1
    LimitParam:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
      example: 20
...
    OrderIDPath:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  responses:
    BadRequest:
      description: Invalid payload, validation error, or invalid identifier/filter.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            default:
              value:
                success: false
                message: validation failed
    Unauthorized:
      description: Missing, invalid, expired, or revoked Bearer token.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            default:
              value:
                success: false
                message: unauthorized
...
    NotFound:
      description: Resource not found.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            default:
              value:
                success: false
                message: not found
    Conflict:
      description: Business conflict, such as duplicate product review.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            default:
              value:
                success: false
                message: review already exists
    TooManyRequests:
      description: Rate limit exceeded.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            default:
              value:
                success: false
                message: too many requests
    InternalServerError:
      description: Unexpected server error. Response does not expose internal SQL/stack details.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            default:
              value:
                success: false
                message: internal server error
...
  schemas:
    ErrorResponse:
      type: object
      required: [success, message]
      properties:
        success:
          type: boolean
          example: false
        message:
          type: string
          example: validation failed
    PaginationMeta:
      type: object
      required: [page, limit, total, total_page, total_pages]
      properties:
        page:
          type: integer
          example: 1
        limit:
          type: integer
          example: 20
        total:
          type: integer
          format: int64
          example: 42
        total_page:
          type: integer
          format: int64
          description: Backward-compatible pagination field.
          example: 3
        total_pages:
          type: integer
          format: int64
          example: 3
...
    OrderProduct:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        image_url:
          type: string
          format: uri
        style:
          type: string
        color:
          type: string
    OrderItem:
      type: object
      properties:
        id:
          type: string
          format: uuid
        product_id:
          type: string
          format: uuid
        quantity:
          type: integer
          minimum: 1
        unit_price:
          type: number
        subtotal:
          type: number
        product:
          $ref: '#/components/schemas/OrderProduct'
    OrderUser:
      type: object
      description: Minimal user info for admin order management. Does not include password hashes or tokens.
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        full_name:
          type: string
        role:
          type: string
          enum: [user, admin]
    Order:
      type: object
      properties:
        id:
          type: string
          format: uuid
        user_id:
          type: string
          format: uuid
        user:
          $ref: '#/components/schemas/OrderUser'
        status:
          type: string
          enum: [pending, paid, shipping, completed, cancelled]
        payment_status:
          type: string
          enum: [unpaid, pending, paid, failed, refunded]
        subtotal_amount:
          type: number
        discount_amount:
          type: number
        coupon_id:
          type: string
          format: uuid
        coupon_code:
          type: string
        total_amount:
          type: number
        recipient_name:
          type: string
        phone:
          type: string
        address_line:
          type: string
        city:
          type: string
        district:
          type: string
        note:
          type: string
        shipping_method:
          type: string
          enum: [standard, express]
        payment_method:
          type: string
          enum: [cod, demo_payment]
        items:
          type: array
          items:
            $ref: '#/components/schemas/OrderItem'
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    CheckoutRequest:
      type: object
      required:
        - recipient_name
        - phone
        - address_line
        - city
        - district
        - shipping_method
        - payment_method
      properties:
        recipient_name:
          type: string
          minLength: 2
          maxLength: 120
        phone:
          type: string
          minLength: 8
          maxLength: 32
        address_line:
          type: string
          minLength: 5
          maxLength: 255
        city:
          type: string
          minLength: 2
          maxLength: 120
        district:
          type: string
          minLength: 2
          maxLength: 120
        note:
          type: string
          maxLength: 1000
        shipping_method:
          type: string
          enum: [standard, express]
        payment_method:
          type: string
          enum: [cod, demo_payment]
        coupon_code:
          type: string
          minLength: 2
          maxLength: 64
...
    OrderResponseEnvelope:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
        data:
          $ref: '#/components/schemas/Order'
    OrderListResponseEnvelope:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
        data:
          type: array
          items:
            $ref: '#/components/schemas/Order'
        meta:
          $ref: '#/components/schemas/PaginationMeta'
    ReturnOrder:
      type: object
      properties:
        id:
          type: string
          format: uuid
        status:
          type: string
          enum: [pending, paid, shipping, completed, cancelled]
        payment_status:
          type: string
          enum: [unpaid, pending, paid, failed, refunded]
        payment_method:
          type: string
          enum: [cod, demo_payment]
        total_amount:
          type: number
        recipient_name:
          type: string
        shipping_method:
          type: string
          enum: [standard, express]
        created_at:
          type: string
          format: date-time
    ReturnUser:
      type: object
      description: Minimal user info for admin return management. Does not include password hashes or tokens.
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        full_name:
          type: string
        role:
          type: string
          enum: [user, admin]
    ReturnRequest:
      type: object
      properties:
        id:
          type: string
          format: uuid
        order_id:
          type: string
          format: uuid
        user_id:
          type: string
          format: uuid
        reason:
          type: string
        status:
          type: string
          enum: [requested, approved, rejected, cancelled]
        admin_note:
          type: string
        order:
          $ref: '#/components/schemas/ReturnOrder'
        user:
          $ref: '#/components/schemas/ReturnUser'
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    CreateReturnRequest:
      type: object
      required: [reason]
      properties:
        reason:
          type: string
          minLength: 10
          maxLength: 1000
...
    ReturnRequestResponseEnvelope:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
        data:
          $ref: '#/components/schemas/ReturnRequest'
    ReturnRequestListResponseEnvelope:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
        data:
          type: array
          items:
            $ref: '#/components/schemas/ReturnRequest'
        meta:
          $ref: '#/components/schemas/PaginationMeta'
...